+++ CHANGELOG.md
@@ -6,6 +6,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ## StreamChat
 ### ✅ Added
 - Add `PinnedMessagesPagination.offset(Int)` for paginating messages using offset [#4249](https://github.com/GetStream/stream-chat-swift/pull/4249)
+- Add `ChatMemberInfo` for slim channel-member info on messages, mentions, and typing events [#4241](https://github.com/GetStream/stream-chat-swift/pull/4241)
 - Add `ChatMessage.mentionedChannelMembers` for accessing channel-member info of mentioned users [#4241](https://github.com/GetStream/stream-chat-swift/pull/4241)
 - Add `TypingUser` and `ChatChannel.typingUsers` for accessing channel-member info of typing users [#4241](https://github.com/GetStream/stream-chat-swift/pull/4241)
 - Add `LivestreamChatState.typingMemberInfos` for accessing channel-member info of typing users [#4241](https://github.com/GetStream/stream-chat-swift/pull/4241)